v1.0 - We've arrived!
Hello! Welcome to v1.0! In my opinion, we've officially reached the point I feel comfortable calling this as good, if not better, than the original. 
Thus, it's the first proper release. There are even multiple features in this version that weren't in the original flash version!
Changes:
You can now see the volume level and difficulty.
The colorful comments of v0.90 have been rolled back.
Happy almost-summer-break any LPS students!